Corrected Grammatical error
In the "If you need Label in your custom View", section, the first point is written as: "create an new view class extends YourView".
This sentence is grammatically incorrect. 'an' is used before a constant which is against the standard grammatical rules.
The corrected sentence could be: "create a new view class extends YourView".
Thank you.